refactor: modularize APV component test suite
- Extract shared test utilities, helper functions, and the `TestListener` class into `TestUtils.h`.
- Relocate YUV and Y4M parameterized encoding tests to `Qv2EncoderTests.cpp`.
- Move component lifecycle, state transition, and factory tests to `Qv2ComponentBaseTests.cpp`.
- Update `CMakeLists.txt` to include the new source files in the `Qv2ComponentTests` build.
- Simplify `testMain.cpp` to focus on the GoogleTest entry point.
